divided plugin.py into data_item
